Enabling support for FLUSH and FUA flags
In order device-mapper target to support FLUSH and FUA requests
one has to set flush_supported flag on the target instance.
This commit does it.

Then dmdedup needs to flush underlying data and metadata devices
when FLUSH and FUA flags are set on a request. This was already done
by the previous code. However, sometimes we see empty write
requests the sole goal of which is to flush the data. This patch
handles this scenario correctly.